+++ src/ape/managers/chain.py
@@ -194,7 +194,7 @@ class BlockContainer(BaseManager):
         """
         Poll new blocks. Optionally set a start block to include historical blocks.
         **NOTE**: This is a daemon method; it does not terminate unless an exception occurrs
-        or a ``stop_block`` is given.
+        or a ``stop`` is given.
 
         Usage example::
 
@@ -207,7 +207,7 @@ class BlockContainer(BaseManager):
             start (Optional[int]): The block number to start with. Defaults to the pending
               block number.
             stop (Optional[int]): Optionally set a future block number to stop at.
-              Defaults to infinity.
+              Defaults to never-ending.
             required_confirmations (Optional[int]): The amount of confirmations to wait
               before yielding the block. The more confirmations, the less likely a reorg will occur.
               Defaults to the network's configured required confirmations.
